Painless Prep/requirements.txt
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
scipy==1.14.1
